% Auto-generated: do not edit by hand
\name{dashAdvancedGrid}

\alias{dashAdvancedGrid}

\title{DashAdvancedGrid component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashAdvancedGrid(id=NULL, data=NULL, header=NULL, rows=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. }

\item{header}{Unnamed list. }

\item{rows}{Unnamed list. }

\item{style}{Named list. }
}
